[AB#33288] Set up frontend to use typescript and modify existing code and tests to use typescript
