+++ private-templates-service/client/package.json
@@ -5,6 +5,10 @@
 	"dependencies": {
 		"@fortawesome/fontawesome-free": "^5.11.2",
 		"@microsoft/microsoft-graph-client": "^2.0.0",
+		"@types/jest": "^24.9.1",
+		"@types/node": "^13.5.0",
+		"@types/react": "^16.9.19",
+		"@types/react-dom": "^16.9.5",
 		"bootstrap": "^4.4.1",
 		"moment": "^2.24.0",
 		"msal": "^1.1.3",
@@ -13,7 +17,8 @@
 		"react-dom": "^16.12.0",
 		"react-router-dom": "^5.1.2",
 		"react-scripts": "3.2.0",
-		"reactstrap": "^8.1.1"
+		"reactstrap": "^8.1.1",
+		"typescript": "^3.7.5"
 	},
 	"scripts": {
 		"start": "react-scripts start",
