--- private-templates-service/client/src/Config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  appId: '4803f66a-136d-4155-a51e-6d98400d5506',
-  redirectUri: 'https://adaptivecms.z22.web.core.windows.net/',
-  scopes: [
-    'user.read',
-  ]
-};
