+++ private-templates-service/client/src/Config.tsx
@@ -1,0 +1,7 @@
+import {} from "react"; // Typescripte requires every file to have an import
+
+export default {
+	appId: "4803f66a-136d-4155-a51e-6d98400d5506",
+	redirectUri: "https://adaptivecms.z22.web.core.windows.net/",
+	scopes: ["user.read"]
+};
