--- private-templates-service/client/src/Welcome.js
@@ -1,40 +1,0 @@
-import React from 'react';
-import {
-  Button,
-  Jumbotron } from 'reactstrap';
-
-function WelcomeContent(props) {
-  // If authenticated, greet the user
-  if (props.isAuthenticated) {
-    if (props.user.organization) {
-      return (
-        <div>
-          <h4>Welcome {props.user.displayName} from {props.user.organization}!</h4>
-        </div>
-      );
-    }
-    return (
-      <div>
-        <h4>Welcome {props.user.displayName}!</h4>
-      </div>
-    );
-  }
-
-  // Not authenticated, present a sign in button
-  return <Button color="primary" onClick={props.authButtonMethod}>Click here to sign in</Button>;
-}
-
-export default class Welcome extends React.Component {
-  render() {
-    return (
-      <Jumbotron>
-        <h1>Admin Portal</h1>
-        <p className="lead">Basic authentication test.</p>
-        <WelcomeContent
-          isAuthenticated={this.props.isAuthenticated}
-          user={this.props.user}
-          authButtonMethod={this.props.authButtonMethod} />
-      </Jumbotron>
-    );
-  }
-}
