+++ private-templates-service/client/src/index.tsx
