+++ private-templates-service/server.js
@@ -6,6 +6,7 @@ const port = process.env.PORT || 5000;
 
 // console.log that your server is up and running
 app.listen(port, () => console.log(`Listening on port ${port}`));
+
 // create a GET route
 app.get("/express_backend", (req, res) => {
 	res.send({
